Finish funcs: _.first through _.reject
Reviewing underbar functions that will be helpful
to utilize in the future. Particularly found
_.reject to be difficult to wrap my head around
more specifically the idea that the 2nd argument
called 'test' should not pass in the test that we
are trying to find the opposite of.  Rather
_.reject takes in an anon function that passes
one item as an argument, runs the test on this
item and returns true if the test on the item
returns false.  Very interesting and clever.