Write functional _.reduce, _.every, and _.some
Found writing a reduce function tremendously
difficult the second time around.
Specifically, trying to understand the use of
_.every within the function.  To review, each
will iterate through an array, and utilize a
callback function on each of the array's
items.  if the accumulator is left undefined
then, the accumulator is set to the first item
in the array, then the reduce will move on to
the next item WITHOUT A RETURN. From that pt
the iterator is set to the accumulator no
matter what the iterator function returns.

Again found much issue trying to resolve
_.every.  Kept forgetting that the callback
function needed to pass in the accumulator
that is set in the parent _.reduce func.

Writing _.some proved to be tremendously fun
given a more thorough understanding of every.
Note to self: Always visually draw out the
logic of a function before attempting to code.