Optimize _.extend, _.defaults, and _.memoize
Review again object helpers and the memoize
function.  Still finding memoize to be difficult
to fully grasp.  Attempted to create a cache by
stringifying the func.apply(this, arguments), but
to no avail.  One possible reason being is that
the apply() method may not be specific enough.

Will review the apply documentation on the
moz://a network.