Develop functional _.delay
Implemented the setTimeout function
to delay the implementation of a function.
Placed a func.apply within an anon func
so that _.delay could be implemented
multiple times in different execution
scopes, which, from my understanding at
the moment, would not be possible if the
func.apply was available to the scope
_.delay is exposed to.  Will have to
look into this at a greater length.